chore: sync master to development

+++ common/helpers.py
@@ -84,7 +84,6 @@ def task_success_callback(context):
 
 
 def process_dynamic_task_sql(sql_query, message):
-
     mongo_filter_field = "patientId"
 
     sql_data = get_data_from_db(db_type="mysql", conn_id="mysql_monolith",
@@ -212,7 +211,7 @@ def get_deactivated_patients():
     :return: None
     """
     _filter = {
-        "userStatus": 3,
+        "userStatus": 5,
         "deleted": False,
         "countryCode": {"$in": [91]},
         "docCode": {"$regex": "^ZH"}
@@ -304,10 +303,11 @@ def add_sales_cm(cm_type):
                                       microsecond=0, tzinfo=local_tz)
     yesterday = today - timedelta(days=1)
     _filter = {
-        "assignedCmType": {"$ne": "sales"},
+        "assignedCmType": "normal",
         "processedSales": {"$ne": True},
         "userStatus": {"$ne": 4},
-        "_created": {"$gt": yesterday}
+        "_created": {"$gt": yesterday},
+        "docCode": {"$regex": "^ZH"}
     }
     eligible_users = get_data_from_db(conn_id="mongo_user_db",
                                       filter=_filter, collection="user")
@@ -391,7 +391,8 @@ def switch_active_cm(cm_type):
             log.info(e)
 
 
-def twilio_cleanup_channel(twilio_service=None, channel_sid=None):
+def twilio_cleanup_channel(twilio_service=None, channel_sid=None,
+                           remove_cm=False):
     """
     This function fetches all members of the target channel defined in
     `channel_sid` and deletes the same from the channel.
@@ -405,7 +406,13 @@ def twilio_cleanup_channel(twilio_service=None, channel_sid=None):
     members = channel.members.list()
     if members:
         for member in members:
-            member.delete()
+            if not remove_cm:
+                member.delete()
+            else:
+                attributes = json.loads(member.attributes)
+                is_cm = attributes.get("isCm", False)
+                if is_cm:
+                    member.delete()
         log.info(channel_sid + " cleaned of all members.")
     else:
         log.info(channel_sid + " has no members to delete.")
@@ -482,20 +489,21 @@ def twilio_cleanup():
                 patient_id) + ". Missing twilio user information")
         try:
             twilio_cleanup_channel(twilio_service=twilio_service,
-                                   channel_sid=channel_sid)
+                                   channel_sid=channel_sid,
+                                   remove_cm=True)
         except Exception as e:
             log.error(e)
-        try:
-            twilio_delete_user(twilio_service=twilio_service,
-                               user_sid=user_sid)
-        except TwilioRestException as e:
-            log.error(e.msg)
-        except Exception as e:
-            log.error(e)
-        try:
-            mark_user_deleted(_id=_id)
-        except Exception as e:
-            log.error(e)
+        # try:
+        #     twilio_delete_user(twilio_service=twilio_service,
+        #                        user_sid=user_sid)
+        # except TwilioRestException as e:
+        #     log.error(e.msg)
+        # except Exception as e:
+        #     log.error(e)
+        # try:
+        #     mark_user_deleted(_id=_id)
+        # except Exception as e:
+        #     log.error(e)
     log.info("Finished processing deactivated users for deletion. ")
 
 
@@ -562,7 +570,8 @@ def refresh_cm_type_user_redis(cm_type="active"):
         _filter = {
             "assignedCmType": cm_type,
             "countryCode": {"$in": [91]},
-            "docCode": {"$regex": doc_code}
+            "docCode": {"$regex": doc_code},
+            "assignedCm": cm
         }
         cacheable_users = get_data_from_db(conn_id="mongo_user_db",
                                            filter=_filter, collection="user")
@@ -920,7 +929,6 @@ def get_created_users_by_cm_by_days(cm_type="sales"):
     users = list(users)
     log.info("old users")
     log.info(len(users))
-    log.info(users)
     return users
 
 
@@ -942,9 +950,7 @@ def continue_statemachine():
                 "docCode": {"$regex": "^ZH"}
             }
             sales_processed_payload = {
-                "processedSales": True,
-                "countryCode": {"$in": [91]},
-                "docCode": {"$regex": "^ZH"}
+                "processedSales": True
             }
             log.info("Fetching user with filter " + json.dumps(remove_filter))
             users = get_data_from_db(
@@ -956,10 +962,9 @@ def continue_statemachine():
                 created_days_users = \
                     get_created_users_by_cm_by_days(
                         cm_type="sales")
-                log.debug(created_days_users)
-                # if created_days_users:
-                #     users = list(users)
-                #     users.extend(created_days_users)
+                if created_days_users:
+                    users = list(users)
+                    users.extend(created_days_users)
             except Exception as e:
                 log.warning(e)
             for user in users:
@@ -992,7 +997,6 @@ def continue_statemachine():
                     method="PATCH")
                 if status == HTTPStatus.OK:
                     log.info("Marked as sales processed. ")
-                    user_list.remove(user_id)
         except Exception as e:
             log.error(e)
             log.error(user_list)
